refactor: implement table and derived at relation
