+++ core/src/data/mod.rs
@@ -2,9 +2,10 @@ mod bigdecimal_ext;
 mod interval;
 mod key;
 mod literal;
+mod relation;
 mod row;
 mod string_ext;
-mod table;
+// mod table;
 
 pub mod schema;
 pub mod value;
@@ -14,9 +15,10 @@ pub use {
     interval::{Interval, IntervalError},
     key::{Key, KeyError},
     literal::{Literal, LiteralError},
+    relation::{get_name, Relation, TableError},
     row::{Row, RowError},
     schema::{Schema, SchemaIndex, SchemaIndexOrd},
     string_ext::{StringExt, StringExtError},
-    table::{get_name, Table, TableError},
+    // table::{get_name, Table, TableError},
     value::{NumericBinaryOperator, Value, ValueError},
 };
