+++ core/src/plan/index.rs
@@ -1,3 +1,5 @@
+use crate::data::Relation;
+
 use {
     crate::{
         ast::{
@@ -69,7 +71,8 @@ fn plan_query(schema_map: &HashMap<String, Schema>, query: Query) -> Result<Quer
     };
 
     let TableWithJoins { relation, .. } = &select.from;
-    let table_name = relation.get_name()?;
+    let relation = Relation::new(relation)?;
+    let table_name = relation.get_name();
     let indexes = match schema_map.get(table_name) {
         Some(Schema { indexes, .. }) => Indexes(indexes.clone()),
         None => {
